# docs/openapi.yaml
openapi: '3.0.3'
info:
  title: NeuroWealth Backend API
  version: '1.0.0'
  description: |
    NeuroWealth autonomous DeFi portfolio manager backend.

    Authentication: All protected endpoints require a `Bearer <JWT>` token in the
    `Authorization` header, obtained from `POST /api/auth/login`.
servers:
  - url: /api/v1
    description: Versioned API base (alias of /api for current release)
  - url: /api
    description: API base
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserWebhookEndpoint:
      type: object
      description: User-scoped outbound webhook endpoint for real-time event delivery.
      required: [id, userId, url, events, isActive, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
          description: Unique endpoint ID.
        userId:
          type: string
          description: Owner of the webhook endpoint.
        url:
          type: string
          format: uri
          description: HTTPS callback URL (must not be localhost or private IP).
        events:
          type: array
          items:
            type: string
          description: Event types to subscribe to (e.g., "transaction.completed", "rebalance.executed").
        topicScope:
          type: string
          nullable: true
          description: Optional topic filter for event scope.
        filterJson:
          type: object
          nullable: true
          description: Optional JSON filter for matching events.
        isActive:
          type: boolean
          description: Whether the endpoint is enabled.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /webhooks/endpoints:
    post:
      operationId: createWebhookEndpoint
      summary: Create a user webhook endpoint
      description: |
        Create a new outbound webhook endpoint for real-time event delivery.
        URL must be HTTPS and not point to localhost or private addresses.
      tags: [Webhooks]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, events]
              properties:
                url:
                  type: string
                  format: uri
                  description: HTTPS callback URL.
                events:
                  type: array
                  items:
                    type: string
                  description: Event types to subscribe to.
                topicScope:
                  type: string
                  nullable: true
                filterJson:
                  type: object
                  nullable: true
      responses:
        '201':
          description: Webhook endpoint created.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/UserWebhookEndpoint'
                  - type: object
                    properties:
                      secret:
                        type: string
                        description: Signing secret (returned once; store securely).
        '400':
          description: Validation failed or invalid URL.
        '401':
          description: Missing/invalid bearer token.

    get:
      operationId: listWebhookEndpoints
      summary: List user webhook endpoints
      tags: [Webhooks]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of webhook endpoints.
          content:
            application/json:
              schema:
                type: object
                properties:
                  endpoints:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserWebhookEndpoint'
        '401':
          description: Missing/invalid bearer token.
  /webhooks/endpoints/{id}:
    get:
      operationId: getWebhookEndpoint
      summary: Get a webhook endpoint
      tags: [Webhooks]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Webhook endpoint ID.
      responses:
        '200':
          description: Webhook endpoint details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWebhookEndpoint'
        '401':
          description: Missing/invalid bearer token.
        '404':
          description: Webhook endpoint not found.

    patch:
      operationId: updateWebhookEndpoint
      summary: Update a webhook endpoint
      tags: [Webhooks]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Webhook endpoint ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    type: string
                topicScope:
                  type: string
                  nullable: true
                filterJson:
                  type: object
                  nullable: true
                isActive:
                  type: boolean
      responses:
        '200':
          description: Webhook endpoint updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWebhookEndpoint'
        '400':
          description: Validation failed.
        '401':
          description: Missing/invalid bearer token.
        '404':
          description: Webhook endpoint not found.

    delete:
      operationId: deleteWebhookEndpoint
      summary: Delete a webhook endpoint
      tags: [Webhooks]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Webhook endpoint ID.
      responses:
        '204':
          description: Webhook endpoint deleted.
        '401':
          description: Missing/invalid bearer token.
        '404':
          description: Webhook endpoint not found.
  /webhooks/endpoints/{id}/rotate-secret:
    post:
      operationId: rotateWebhookSecret
      summary: Rotate webhook endpoint secret
      description: |
        Begin secret rotation with dual-signing period. Returns new secret and
        activation time. Old secret remains valid until activation time.
      tags: [Webhooks]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Webhook endpoint ID.
      responses:
        '200':
          description: Secret rotation initiated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  secretNext:
                    type: string
                    description: New secret (store securely).
                  secretNextActiveAt:
                    type: string
                    format: date-time
                    description: When new secret becomes active.
                  message:
                    type: string
        '401':
          description: Missing/invalid bearer token.
        '404':
          description: Webhook endpoint not found.
  /webhooks/endpoints/{id}/test:
    post:
      operationId: testWebhookEndpoint
      summary: Send a test ping to webhook endpoint
      description: Sends a test event to verify endpoint is reachable and signature validation works.
      tags: [Webhooks]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Webhook endpoint ID.
      responses:
        '200':
          description: Test event sent.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        '401':
          description: Missing/invalid bearer token.
        '404':
          description: Webhook endpoint not found.
...